+++ packages/core/src/index.ts
@@ -13,6 +13,8 @@ export { BestiaryContext, useBestiary } from "./BestiaryContext";
 export type { BestiaryContextValue, BestiaryCreatureRef } from "./BestiaryContext";
 export { NpcContext, useNpcs } from "./NpcContext";
 export type { NpcContextValue, NpcRef } from "./NpcContext";
+export { XpContext, useXp } from "./XpContext";
+export type { XpContextValue } from "./XpContext";
 export { CalendarContext, useCalendar } from "./CalendarContext";
 export type { CalendarContextValue } from "./CalendarContext";
 export type {